move UserView into User folder
